fix: all test updated and objects have variables
