+++ server/apiResponse.js
@@ -1,0 +1,125 @@
+const apiResponse = [
+  {
+    id: 343935,
+    name: 'Lloyd Austin III',
+    title: 'Secretary of Defense',
+    address_line1: 'U.S. Department of Defense',
+    address_line2: '1000 Defense Pentagon',
+    address_city: 'Washington',
+    address_state: 'DC',
+    address_zip: '20301-1000',
+    address_country: 'US',
+    email: 'Not Made Public',
+    contactPage:
+      'https://www.defense.gov/our-story/meet-the-team/secretary-of-defense/',
+    photoUrl:
+      'https://media.defense.gov/2021/Jan/29/2002592436/825/780/0/210123-A-CN110-0001.JPG',
+    socialMediaPages: [
+      {
+        type: 'facebook',
+        url: 'https://facebook.com/DeptofDefense',
+        icon: 'fa-brands fa-facebook-f',
+        color: '#4267B2'
+      },
+      {
+        type: 'instagram',
+        url: 'https://instagram.com/austin_lloy',
+        icon: 'fa-brands fa-instagram',
+        color: '#C13584'
+      },
+      {
+        type: 'twitter',
+        url: 'https://twitter.com/SecDef',
+        icon: 'fa-brands fa-twitter',
+        color: '#1DA1F2'
+      },
+      {
+        type: 'twitter',
+        url: 'https://twitter.com/DeptofDefense',
+        icon: 'fa-brands fa-twitter',
+        color: '#1DA1F2'
+      }
+    ],
+    photoCroppingCSS: 'center center'
+  },
+  {
+    id: 6641,
+    name: 'Xavier Becerra',
+    title: 'Secretary of Health and Human Services',
+    address_line1: 'U.S. Department of Health and Human Services',
+    address_line2: '200 Independence Avenue, SW',
+    address_city: 'Washington',
+    address_state: 'DC',
+    address_zip: '20201',
+    address_country: 'US',
+    email: 'Not Made Public',
+    contactPage: 'https://www.hhs.gov/about/leadership/xavier-becerra.html',
+    photoUrl:
+      'https://www.hhs.gov/sites/default/files/styles/bio_280x280/public/secretary-xavier-becerra.jpg?h=a0d58e75&itok=dCzPbiYp',
+    socialMediaPages: [
+      {
+        type: 'twitter',
+        url: 'https://twitter.com/SecBecerra',
+        icon: 'fa-brands fa-twitter',
+        color: '#1DA1F2'
+      },
+      {
+        type: 'twitter',
+        url: 'https://twitter.com/XavierBecerra',
+        icon: 'fa-brands fa-twitter',
+        color: '#1DA1F2'
+      }
+    ],
+    photoCroppingCSS: 'center center'
+  },
+  {
+    id: 343167,
+    name: 'Tony Blinken',
+    title: 'Secretary of State',
+    address_line1: 'Harry S. Truman Building',
+    address_line2: '2201 C Street, NW',
+    address_city: 'Washington',
+    address_state: 'DC',
+    address_zip: '20520',
+    address_country: 'US',
+    email: 'Not Made Public',
+    contactPage: 'https://register.state.gov/contactus/contactusform',
+    photoUrl:
+      'https://www.state.gov/wp-content/uploads/2021/01/50925278638_389ad88a3f_k-1125x1406.jpg',
+    socialMediaPages: [
+      {
+        type: 'facebook',
+        url: 'https://facebook.com/statedept',
+        icon: 'fa-brands fa-facebook-f',
+        color: '#4267B2'
+      },
+      {
+        type: 'instagram',
+        url: 'https://instagram.com/statedept',
+        icon: 'fa-brands fa-instagram',
+        color: '#C13584'
+      },
+      {
+        type: 'instagram',
+        url: 'https://instagram.com/secblinken',
+        icon: 'fa-brands fa-instagram',
+        color: '#C13584'
+      },
+      {
+        type: 'twitter',
+        url: 'https://twitter.com/SecBlinken',
+        icon: 'fa-brands fa-twitter',
+        color: '#1DA1F2'
+      },
+      {
+        type: 'youtube',
+        url: 'https://youtube.com/statevideo',
+        icon: 'fa-brands fa-youtube',
+        color: '#FF0000'
+      }
+    ],
+    photoCroppingCSS: 'center center'
+  }
+]
+
+module.exports = apiResponse
